+++ Rpkg/metropolis/R/data.R
@@ -1,4 +1,4 @@
-#' A case control study of childhood leukemia and magnetic fields from Savitz, Wachtel, Barnes, et al (1998) <doi:10.1093/oxfordjournals.aje.a114943>
+#' A case control study of childhood leukemia and magnetic fields from Savitz, Wachtel, Barnes, et al (1998) \doi{doi:10.1093/oxfordjournals.aje.a114943}.
 #'
 #' @format A data frame with 234 rows and 2 variables:
 #' \describe{